refactor: split needs_human retirement helpers out of operator_hints.py
The first-party 1500-line maintainability guardrail
(tests/unit/test_core_decomposition_maintainability.py) flagged
src/awf/runtime/pr_monitor_runner/operator_hints.py after the base sync.
Move the two self-contained monitor-state bookkeeping helpers —
_finalize_processed_operator_hint and
_mark_referenced_needs_human_feedback_answered — into a sibling
operator_hint_retirement module and re-export them with `X as X` so the
existing import sites (and the tests that import them from
operator_hints) keep working unchanged.

Pure code motion: no behavior change. operator_hints.py drops from 1405
to 1299 lines, leaving headroom under the 1350-line target for later
review fixes.
